+++ saleor/graphql/checkout/tests/test_checkout.py
@@ -1,7 +1,6 @@
 import datetime
 from decimal import Decimal
 from unittest import mock
-from unittest.mock import patch
 
 import graphene
 import pytest
@@ -14,28 +13,18 @@ from django_countries.fields import Country
 from measurement.measures import Weight
 from prices import Money
 
-from ....account.models import User
 from ....checkout import base_calculations, calculations
 from ....checkout.checkout_cleaner import (
     clean_checkout_payment,
     clean_checkout_shipping,
 )
 from ....checkout.error_codes import CheckoutErrorCode
-from ....checkout.fetch import (
-    fetch_checkout_info,
-    fetch_checkout_lines,
-    get_delivery_method_info,
-)
-from ....checkout.utils import PRIVATE_META_APP_SHIPPING_ID
-from ....core.payments import PaymentInterface
+from ....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
 from ....core.prices import quantize_price
-from ....payment import TransactionAction, TransactionKind
-from ....payment.interface import GatewayResponse
-from ....plugins.base_plugin import ExcludedShippingMethod
+from ....payment import TransactionAction
 from ....plugins.manager import get_plugins_manager
 from ....plugins.tests.sample_plugins import ActiveDummyPaymentGateway
 from ....product.models import ProductVariant
-from ....shipping import models as shipping_models
 from ....shipping.models import ShippingMethodTranslation
 from ....shipping.utils import convert_to_shipping_method_data
 from ....tests.utils import dummy_editorjs
@@ -878,184 +867,6 @@ def test_checkout_avail_collect_fallbacks_to_channel_country_when_no_shipping_ad
     ]
 
 
-MUTATION_CHECKOUT_CUSTOMER_ATTACH = """
-    mutation checkoutCustomerAttach($id: ID, $customerId: ID) {
-        checkoutCustomerAttach(id: $id, customerId: $customerId) {
-            checkout {
-                token
-            }
-            errors {
-                code
-                field
-                message
-            }
-        }
-    }
-"""
-
-
-def test_checkout_customer_attach(
-    user_api_client, checkout_with_item, customer_user, permission_impersonate_user
-):
-    checkout = checkout_with_item
-    checkout.email = "old@email.com"
-    checkout.save()
-    assert checkout.user is None
-    previous_last_change = checkout.last_change
-
-    query = MUTATION_CHECKOUT_CUSTOMER_ATTACH
-    customer_id = graphene.Node.to_global_id("User", customer_user.pk)
-    variables = {"id": to_global_id_or_none(checkout), "customerId": customer_id}
-
-    response = user_api_client.post_graphql(
-        query, variables, permissions=[permission_impersonate_user]
-    )
-    content = get_graphql_content(response)
-
-    data = content["data"]["checkoutCustomerAttach"]
-    assert not data["errors"]
-    checkout.refresh_from_db()
-    assert checkout.user == customer_user
-    assert checkout.email == customer_user.email
-    assert checkout.last_change != previous_last_change
-
-
-def test_checkout_customer_attach_no_customer_id(
-    api_client, user_api_client, checkout_with_item, customer_user
-):
-    checkout = checkout_with_item
-    checkout.email = "old@email.com"
-    checkout.save()
-    assert checkout.user is None
-    previous_last_change = checkout.last_change
-
-    query = MUTATION_CHECKOUT_CUSTOMER_ATTACH
-    variables = {"id": to_global_id_or_none(checkout)}
-
-    # Mutation should fail for unauthenticated customers
-    response = api_client.post_graphql(query, variables)
-    assert_no_permission(response)
-
-    # Mutation should succeed for authenticated customer
-    response = user_api_client.post_graphql(query, variables)
-    content = get_graphql_content(response)
-    data = content["data"]["checkoutCustomerAttach"]
-    assert not data["errors"]
-    checkout.refresh_from_db()
-    assert checkout.user == customer_user
-    assert checkout.email == customer_user.email
-    assert checkout.last_change != previous_last_change
-
-
-def test_checkout_customer_attach_by_app(
-    app_api_client, checkout_with_item, customer_user, permission_impersonate_user
-):
-    checkout = checkout_with_item
-    checkout.email = "old@email.com"
-    checkout.save()
-    assert checkout.user is None
-    previous_last_change = checkout.last_change
-
-    query = MUTATION_CHECKOUT_CUSTOMER_ATTACH
-    customer_id = graphene.Node.to_global_id("User", customer_user.pk)
-    variables = {"id": to_global_id_or_none(checkout), "customerId": customer_id}
-
-    # Mutation should succeed for authenticated customer
-    response = app_api_client.post_graphql(
-        query, variables, permissions=[permission_impersonate_user]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["checkoutCustomerAttach"]
-    assert not data["errors"]
-    checkout.refresh_from_db()
-    assert checkout.user == customer_user
-    assert checkout.email == customer_user.email
-    assert checkout.last_change != previous_last_change
-
-
-def test_checkout_customer_attach_by_app_no_customer_id(
-    app_api_client, checkout_with_item, permission_impersonate_user
-):
-    checkout = checkout_with_item
-    checkout.email = "old@email.com"
-    checkout.save()
-    assert checkout.user is None
-
-    query = MUTATION_CHECKOUT_CUSTOMER_ATTACH
-    variables = {"id": to_global_id_or_none(checkout)}
-
-    # Mutation should succeed for authenticated customer
-    response = app_api_client.post_graphql(
-        query,
-        variables,
-        permissions=[permission_impersonate_user],
-        check_no_permissions=False,
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["checkoutCustomerAttach"]
-    assert len(data["errors"]) == 1
-    assert data["errors"][0]["code"] == CheckoutErrorCode.REQUIRED.name
-    assert data["errors"][0]["field"] == "customerId"
-
-
-def test_checkout_customer_attach_by_app_without_permission(
-    app_api_client, checkout_with_item, customer_user
-):
-    checkout = checkout_with_item
-    checkout.email = "old@email.com"
-    checkout.save()
-    assert checkout.user is None
-
-    query = MUTATION_CHECKOUT_CUSTOMER_ATTACH
-    customer_id = graphene.Node.to_global_id("User", customer_user.pk)
-    variables = {"id": to_global_id_or_none(checkout), "customerId": customer_id}
-
-    # Mutation should succeed for authenticated customer
-    response = app_api_client.post_graphql(
-        query,
-        variables,
-    )
-
-    assert_no_permission(response)
-
-
-def test_checkout_customer_attach_user_to_checkout_with_user(
-    user_api_client, customer_user, user_checkout, address
-):
-    checkout = user_checkout
-
-    query = """
-    mutation checkoutCustomerAttach($id: ID) {
-        checkoutCustomerAttach(id: $id) {
-            checkout {
-                token
-            }
-            errors {
-                field
-                message
-                code
-            }
-        }
-    }
-"""
-
-    default_address = address.get_copy()
-    second_user = User.objects.create_user(
-        "test2@example.com",
-        "password",
-        default_billing_address=default_address,
-        default_shipping_address=default_address,
-        first_name="Test2",
-        last_name="Tested",
-    )
-
-    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
-    customer_id = graphene.Node.to_global_id("User", second_user.pk)
-    variables = {"id": checkout_id, "customerId": customer_id}
-    response = user_api_client.post_graphql(query, variables)
-    assert_no_permission(response)
-
-
 GET_CHECKOUT_STOCK_RESERVATION_EXPIRES_QUERY = """
 query getCheckoutStockReservationExpiration($id: ID) {
     checkout(id: $id) {
@@ -1195,173 +1006,21 @@ def test_checkout_reservation_date_for_disabled_reservations(
     assert content["data"]["checkout"]["stockReservationExpires"] is None
 
 
-MUTATION_CHECKOUT_CUSTOMER_DETACH = """
-    mutation checkoutCustomerDetach($id: ID) {
-        checkoutCustomerDetach(id: $id) {
-            checkout {
-                token
-            }
-            errors {
-                field
-                message
-            }
-        }
-    }
-    """
+# @pytest.fixture
+# def fake_manager(mocker):
+#     return mocker.Mock(spec=PaymentInterface)
 
 
-def test_checkout_customer_detach(user_api_client, checkout_with_item, customer_user):
-    checkout = checkout_with_item
-    checkout.user = customer_user
-    checkout.save(update_fields=["user"])
-    previous_last_change = checkout.last_change
+# @pytest.fixture
+# def mock_get_manager(mocker, fake_manager):
+#     manager = mocker.patch(
+#         "saleor.payment.gateway.get_plugins_manager",
+#         autospec=True,
+#         return_value=fake_manager,
+#     )
+#     yield fake_manager
+#     manager.assert_called_once()
 
-    variables = {"id": to_global_id_or_none(checkout)}
-
-    # Mutation should succeed if the user owns this checkout.
-    response = user_api_client.post_graphql(
-        MUTATION_CHECKOUT_CUSTOMER_DETACH, variables
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["checkoutCustomerDetach"]
-    assert not data["errors"]
-    checkout.refresh_from_db()
-    assert checkout.user is None
-    assert checkout.last_change != previous_last_change
-
-    # Mutation should fail when user calling it doesn't own the checkout.
-    other_user = User.objects.create_user("othercustomer@example.com", "password")
-    checkout.user = other_user
-    checkout.save()
-    response = user_api_client.post_graphql(
-        MUTATION_CHECKOUT_CUSTOMER_DETACH, variables
-    )
-    assert_no_permission(response)
-
-
-def test_checkout_customer_detach_by_app(
-    app_api_client, checkout_with_item, customer_user, permission_impersonate_user
-):
-    checkout = checkout_with_item
-    checkout.user = customer_user
-    checkout.save(update_fields=["user"])
-    previous_last_change = checkout.last_change
-
-    variables = {"id": to_global_id_or_none(checkout)}
-
-    # Mutation should succeed if the user owns this checkout.
-    response = app_api_client.post_graphql(
-        MUTATION_CHECKOUT_CUSTOMER_DETACH,
-        variables,
-        permissions=[permission_impersonate_user],
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["checkoutCustomerDetach"]
-    assert not data["errors"]
-    checkout.refresh_from_db()
-    assert checkout.user is None
-    assert checkout.last_change != previous_last_change
-
-
-def test_checkout_customer_detach_by_app_without_permissions(
-    app_api_client, checkout_with_item, customer_user
-):
-    checkout = checkout_with_item
-    checkout.user = customer_user
-    checkout.save(update_fields=["user"])
-    previous_last_change = checkout.last_change
-
-    variables = {"id": to_global_id_or_none(checkout)}
-
-    # Mutation should succeed if the user owns this checkout.
-    response = app_api_client.post_graphql(MUTATION_CHECKOUT_CUSTOMER_DETACH, variables)
-
-    assert_no_permission(response)
-    checkout.refresh_from_db()
-    assert checkout.last_change == previous_last_change
-
-
-CHECKOUT_EMAIL_UPDATE_MUTATION = """
-    mutation checkoutEmailUpdate($id: ID, $email: String!) {
-        checkoutEmailUpdate(id: $id, email: $email) {
-            checkout {
-                id,
-                email
-            },
-            errors {
-                field,
-                message
-            }
-            errors {
-                field,
-                message
-                code
-            }
-        }
-    }
-"""
-
-
-def test_checkout_email_update(user_api_client, checkout_with_item):
-    checkout = checkout_with_item
-    checkout.email = None
-    checkout.save(update_fields=["email"])
-    previous_last_change = checkout.last_change
-
-    email = "test@example.com"
-    variables = {"id": to_global_id_or_none(checkout), "email": email}
-
-    response = user_api_client.post_graphql(CHECKOUT_EMAIL_UPDATE_MUTATION, variables)
-    content = get_graphql_content(response)
-    data = content["data"]["checkoutEmailUpdate"]
-    assert not data["errors"]
-    checkout.refresh_from_db()
-    assert checkout.email == email
-    assert checkout.last_change != previous_last_change
-
-
-def test_checkout_email_update_validation(user_api_client, checkout_with_item):
-    variables = {"id": to_global_id_or_none(checkout_with_item), "email": ""}
-
-    response = user_api_client.post_graphql(CHECKOUT_EMAIL_UPDATE_MUTATION, variables)
-    content = get_graphql_content(response)
-    previous_last_change = checkout_with_item.last_change
-
-    errors = content["data"]["checkoutEmailUpdate"]["errors"]
-    assert errors
-    assert errors[0]["field"] == "email"
-    assert errors[0]["message"] == "This field cannot be blank."
-
-    checkout_errors = content["data"]["checkoutEmailUpdate"]["errors"]
-    assert checkout_errors[0]["code"] == CheckoutErrorCode.REQUIRED.name
-    assert checkout_with_item.last_change == previous_last_change
-
-
-@pytest.fixture
-def fake_manager(mocker):
-    return mocker.Mock(spec=PaymentInterface)
-
-
-@pytest.fixture
-def mock_get_manager(mocker, fake_manager):
-    manager = mocker.patch(
-        "saleor.payment.gateway.get_plugins_manager",
-        autospec=True,
-        return_value=fake_manager,
-    )
-    yield fake_manager
-    manager.assert_called_once()
-
-
-TRANSACTION_CONFIRM_GATEWAY_RESPONSE = GatewayResponse(
-    is_success=False,
-    action_required=False,
-    kind=TransactionKind.CONFIRM,
-    amount=Decimal(3.0),
-    currency="usd",
-    transaction_id="1234",
-    error=None,
-)
 
 QUERY_CHECKOUT_USER_ID = """
     query getCheckout($id: ID) {
@@ -2011,688 +1670,6 @@ def test_checkout_prices_with_voucher_once_per_order(
         data["lines"][0]["undiscountedTotalPrice"]["amount"]
         == undiscounted_unit_price.amount * line_info.line.quantity
     )
-
-
-MUTATION_UPDATE_SHIPPING_METHOD = """
-    mutation checkoutShippingMethodUpdate(
-            $id: ID, $shippingMethodId: ID!){
-        checkoutShippingMethodUpdate(
-            id: $id, shippingMethodId: $shippingMethodId) {
-            errors {
-                field
-                message
-                code
-            }
-            checkout {
-                token
-            }
-        }
-    }
-"""
-
-MUTATION_UPDATE_DELIVERY_METHOD = """
-    mutation checkoutDeliveryMethodUpdate(
-            $id: ID, $deliveryMethodId: ID) {
-        checkoutDeliveryMethodUpdate(
-            id: $id,
-            deliveryMethodId: $deliveryMethodId) {
-            checkout {
-            id
-            deliveryMethod {
-                __typename
-                ... on ShippingMethod {
-                    name
-                    id
-                    translation(languageCode: EN_US) {
-                        name
-                    }
-                }
-                ... on Warehouse {
-                   name
-                   id
-                }
-            }
-        }
-        errors {
-            field
-            message
-            code
-        }
-    }
-}
-"""
-
-
-# TODO: Deprecated
-@pytest.mark.parametrize("is_valid_shipping_method", (True, False))
-@patch(
-    "saleor.graphql.checkout.mutations.checkout_shipping_method_update."
-    "clean_delivery_method"
-)
-def test_checkout_shipping_method_update(
-    mock_clean_shipping,
-    staff_api_client,
-    shipping_method,
-    checkout_with_item_and_shipping_method,
-    is_valid_shipping_method,
-):
-    checkout = checkout_with_item_and_shipping_method
-    old_shipping_method = checkout.shipping_method
-    query = MUTATION_UPDATE_SHIPPING_METHOD
-    mock_clean_shipping.return_value = is_valid_shipping_method
-    previous_last_change = checkout.last_change
-
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-
-    response = staff_api_client.post_graphql(
-        query, {"id": to_global_id_or_none(checkout), "shippingMethodId": method_id}
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-    checkout.refresh_from_db()
-
-    manager = get_plugins_manager()
-    lines, _ = fetch_checkout_lines(checkout)
-    checkout_info = fetch_checkout_info(checkout, lines, [], manager)
-    checkout_info.delivery_method_info = get_delivery_method_info(
-        convert_to_shipping_method_data(
-            old_shipping_method, old_shipping_method.channel_listings.first()
-        ),
-        None,
-    )
-    mock_clean_shipping.assert_called_once_with(
-        checkout_info=checkout_info,
-        lines=lines,
-        method=convert_to_shipping_method_data(
-            shipping_method, shipping_method.channel_listings.first()
-        ),
-    )
-    errors = data["errors"]
-    if is_valid_shipping_method:
-        assert not errors
-        assert data["checkout"]["token"] == str(checkout.token)
-        assert checkout.shipping_method == shipping_method
-        assert checkout.last_change != previous_last_change
-    else:
-        assert len(errors) == 1
-        assert errors[0]["field"] == "shippingMethod"
-        assert (
-            errors[0]["code"] == CheckoutErrorCode.SHIPPING_METHOD_NOT_APPLICABLE.name
-        )
-        assert checkout.shipping_method == old_shipping_method
-        assert checkout.last_change == previous_last_change
-
-
-@mock.patch("saleor.plugins.webhook.tasks.send_webhook_request_sync")
-def test_checkout_shipping_method_update_external_shipping_method(
-    mock_send_request,
-    staff_api_client,
-    address,
-    checkout_with_item,
-    shipping_app,
-    channel_USD,
-    settings,
-):
-    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
-    response_method_id = "abcd"
-    mock_json_response = [
-        {
-            "id": response_method_id,
-            "name": "Provider - Economy",
-            "amount": "10",
-            "currency": "USD",
-            "maximum_delivery_days": "7",
-        }
-    ]
-    mock_send_request.return_value = mock_json_response
-
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-
-    method_id = graphene.Node.to_global_id(
-        "app", f"{shipping_app.id}:{response_method_id}"
-    )
-
-    response = staff_api_client.post_graphql(
-        MUTATION_UPDATE_SHIPPING_METHOD,
-        {"id": to_global_id_or_none(checkout_with_item), "shippingMethodId": method_id},
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-    checkout.refresh_from_db()
-
-    errors = data["errors"]
-    assert not errors
-    assert data["checkout"]["token"] == str(checkout_with_item.token)
-    assert PRIVATE_META_APP_SHIPPING_ID in checkout.private_metadata
-
-
-@mock.patch("saleor.plugins.webhook.tasks.send_webhook_request_sync")
-def test_checkout_shipping_method_update_external_shipping_method_with_tax_plugin(
-    mock_send_request,
-    staff_api_client,
-    address,
-    checkout_with_item,
-    shipping_app,
-    channel_USD,
-    settings,
-):
-    settings.PLUGINS = [
-        "saleor.plugins.tests.sample_plugins.PluginSample",
-        "saleor.plugins.webhook.plugin.WebhookPlugin",
-    ]
-    response_method_id = "abcd"
-    mock_json_response = [
-        {
-            "id": response_method_id,
-            "name": "Provider - Economy",
-            "amount": "10",
-            "currency": "USD",
-            "maximum_delivery_days": "7",
-        }
-    ]
-    mock_send_request.return_value = mock_json_response
-
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save()
-
-    method_id = graphene.Node.to_global_id(
-        "app", f"{shipping_app.id}:{response_method_id}"
-    )
-
-    # Set external shipping method for first time
-    response = staff_api_client.post_graphql(
-        MUTATION_UPDATE_SHIPPING_METHOD,
-        {"id": to_global_id_or_none(checkout_with_item), "shippingMethodId": method_id},
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-    assert not data["errors"]
-
-    # Set external shipping for second time
-    # Without a fix this request results in infinite recursion
-    # between Avalara and Webhooks plugins
-    response = staff_api_client.post_graphql(
-        MUTATION_UPDATE_SHIPPING_METHOD,
-        {"id": to_global_id_or_none(checkout_with_item), "shippingMethodId": method_id},
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-    assert not data["errors"]
-
-
-@pytest.mark.parametrize("is_valid_delivery_method", (True, False))
-@pytest.mark.parametrize(
-    "delivery_method, node_name, attribute_name",
-    [
-        ("warehouse", "Warehouse", "collection_point"),
-        ("shipping_method", "ShippingMethod", "shipping_method"),
-    ],
-    indirect=("delivery_method",),
-)
-@patch(
-    "saleor.graphql.checkout.mutations.checkout_delivery_method_update."
-    "clean_delivery_method"
-)
-def test_checkout_delivery_method_update(
-    mock_clean_delivery,
-    api_client,
-    delivery_method,
-    node_name,
-    attribute_name,
-    checkout_with_item_for_cc,
-    is_valid_delivery_method,
-):
-    # given
-    mock_clean_delivery.return_value = is_valid_delivery_method
-
-    checkout = checkout_with_item_for_cc
-    manager = get_plugins_manager()
-    lines, _ = fetch_checkout_lines(checkout)
-    checkout_info = fetch_checkout_info(checkout, lines, [], manager)
-
-    shipping_method_data = delivery_method
-    if attribute_name == "shipping_method":
-        shipping_method_data = convert_to_shipping_method_data(
-            delivery_method,
-            delivery_method.channel_listings.get(),
-        )
-    query = MUTATION_UPDATE_DELIVERY_METHOD
-    mock_clean_delivery.return_value = is_valid_delivery_method
-
-    method_id = graphene.Node.to_global_id(node_name, delivery_method.id)
-
-    response = api_client.post_graphql(
-        query, {"id": to_global_id_or_none(checkout), "deliveryMethodId": method_id}
-    )
-    data = get_graphql_content(response)["data"]["checkoutDeliveryMethodUpdate"]
-    checkout.refresh_from_db()
-
-    mock_clean_delivery.assert_called_once_with(
-        checkout_info=checkout_info, lines=lines, method=shipping_method_data
-    )
-    errors = data["errors"]
-    if is_valid_delivery_method:
-        assert not errors
-        assert getattr(checkout, attribute_name) == delivery_method
-    else:
-        assert len(errors) == 1
-        assert errors[0]["field"] == "deliveryMethodId"
-        assert (
-            errors[0]["code"] == CheckoutErrorCode.DELIVERY_METHOD_NOT_APPLICABLE.name
-        )
-        assert checkout.shipping_method is None
-        assert checkout.collection_point is None
-
-
-@pytest.mark.parametrize("is_valid_delivery_method", (True, False))
-@mock.patch("saleor.plugins.webhook.tasks.send_webhook_request_sync")
-@patch(
-    "saleor.graphql.checkout.mutations.checkout_delivery_method_update."
-    "clean_delivery_method"
-)
-def test_checkout_delivery_method_update_external_shipping(
-    mock_clean_delivery,
-    mock_send_request,
-    api_client,
-    checkout_with_item_for_cc,
-    is_valid_delivery_method,
-    settings,
-    shipping_app,
-    channel_USD,
-):
-    checkout = checkout_with_item_for_cc
-    query = MUTATION_UPDATE_DELIVERY_METHOD
-    mock_clean_delivery.return_value = is_valid_delivery_method
-
-    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
-    response_method_id = "abcd"
-    mock_json_response = [
-        {
-            "id": response_method_id,
-            "name": "Provider - Economy",
-            "amount": "10",
-            "currency": "USD",
-            "maximum_delivery_days": "7",
-        }
-    ]
-    mock_send_request.return_value = mock_json_response
-
-    method_id = graphene.Node.to_global_id(
-        "app", f"{shipping_app.id}:{response_method_id}"
-    )
-
-    response = api_client.post_graphql(
-        query, {"id": to_global_id_or_none(checkout), "deliveryMethodId": method_id}
-    )
-    data = get_graphql_content(response)["data"]["checkoutDeliveryMethodUpdate"]
-    checkout.refresh_from_db()
-
-    errors = data["errors"]
-    if is_valid_delivery_method:
-        assert not errors
-        assert PRIVATE_META_APP_SHIPPING_ID in checkout.private_metadata
-        assert data["checkout"]["deliveryMethod"]["id"] == method_id
-    else:
-        assert len(errors) == 1
-        assert errors[0]["field"] == "deliveryMethodId"
-        assert (
-            errors[0]["code"] == CheckoutErrorCode.DELIVERY_METHOD_NOT_APPLICABLE.name
-        )
-        assert PRIVATE_META_APP_SHIPPING_ID not in checkout.private_metadata
-
-
-@patch(
-    "saleor.graphql.checkout.mutations.checkout_shipping_method_update."
-    "clean_delivery_method"
-)
-def test_checkout_delivery_method_update_with_id_of_different_type_causes_and_error(
-    mock_clean_delivery,
-    api_client,
-    checkout_with_item,
-    address,
-):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_DELIVERY_METHOD
-    mock_clean_delivery.return_value = True
-    invalid_method_id = graphene.Node.to_global_id("Address", address.id)
-
-    response = api_client.post_graphql(
-        query,
-        {
-            "id": to_global_id_or_none(checkout),
-            "deliveryMethodId": invalid_method_id,
-        },
-    )
-    data = get_graphql_content(response)["data"]["checkoutDeliveryMethodUpdate"]
-    checkout.refresh_from_db()
-
-    errors = data["errors"]
-    assert len(errors) == 1
-    assert errors[0]["field"] == "deliveryMethodId"
-    assert errors[0]["code"] == CheckoutErrorCode.INVALID.name
-    assert checkout.shipping_method is None
-    assert checkout.collection_point is None
-
-
-@patch(
-    "saleor.graphql.checkout.mutations.checkout_shipping_method_update."
-    "clean_delivery_method"
-)
-def test_checkout_delivery_method_with_nonexistant_id_results_not_found(
-    mock_clean_delivery,
-    api_client,
-    warehouse_for_cc,
-    checkout_with_item,
-    address,
-):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_DELIVERY_METHOD
-    mock_clean_delivery.return_value = True
-
-    nonexistant_id = "YXBwOjEyMzQ6c29tZS1pZA=="
-    response = api_client.post_graphql(
-        query,
-        {
-            "id": to_global_id_or_none(checkout),
-            "deliveryMethodId": nonexistant_id,
-        },
-    )
-    data = get_graphql_content(response)["data"]["checkoutDeliveryMethodUpdate"]
-    checkout.refresh_from_db()
-
-    assert not data["checkout"]
-    assert data["errors"][0]["field"] == "deliveryMethodId"
-    assert data["errors"][0]["code"] == CheckoutErrorCode.NOT_FOUND.name
-    assert checkout.shipping_method is None
-    assert checkout.collection_point is None
-
-
-@patch(
-    "saleor.graphql.checkout.mutations.checkout_shipping_method_update."
-    "clean_delivery_method"
-)
-def test_checkout_delivery_method_with_empty_fields_results_None(
-    mock_clean_delivery,
-    api_client,
-    warehouse_for_cc,
-    checkout_with_item,
-    address,
-):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_DELIVERY_METHOD
-    mock_clean_delivery.return_value = True
-
-    response = api_client.post_graphql(
-        query,
-        {
-            "id": to_global_id_or_none(checkout),
-        },
-    )
-    data = get_graphql_content(response)["data"]["checkoutDeliveryMethodUpdate"]
-    checkout.refresh_from_db()
-
-    assert not data["errors"]
-    assert data["checkout"]["deliveryMethod"] is None
-    assert checkout.shipping_method is None
-    assert checkout.collection_point is None
-
-
-@mock.patch(
-    "saleor.plugins.manager.PluginsManager.excluded_shipping_methods_for_checkout"
-)
-def test_checkout_shipping_method_update_excluded_webhook(
-    mocked_webhook,
-    staff_api_client,
-    shipping_method,
-    checkout_with_item,
-    address,
-):
-    # given
-    webhook_reason = "spanish-inquisition"
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_SHIPPING_METHOD
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-    mocked_webhook.return_value = [
-        ExcludedShippingMethod(shipping_method.id, webhook_reason)
-    ]
-
-    # when
-    response = staff_api_client.post_graphql(
-        query,
-        {"id": to_global_id_or_none(checkout_with_item), "shippingMethodId": method_id},
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-
-    checkout.refresh_from_db()
-
-    # then
-    errors = data["errors"]
-    assert len(errors) == 1
-    assert errors[0]["field"] == "shippingMethod"
-    assert errors[0]["code"] == CheckoutErrorCode.SHIPPING_METHOD_NOT_APPLICABLE.name
-    assert checkout.shipping_method is None
-
-
-# Deprecated
-@patch("saleor.shipping.postal_codes.is_shipping_method_applicable_for_postal_code")
-def test_checkout_shipping_method_update_excluded_postal_code(
-    mock_is_shipping_method_available,
-    staff_api_client,
-    shipping_method,
-    checkout_with_item,
-    address,
-):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_SHIPPING_METHOD
-    mock_is_shipping_method_available.return_value = False
-
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-
-    response = staff_api_client.post_graphql(
-        query,
-        {"id": to_global_id_or_none(checkout_with_item), "shippingMethodId": method_id},
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-
-    checkout.refresh_from_db()
-
-    errors = data["errors"]
-    assert len(errors) == 1
-    assert errors[0]["field"] == "shippingMethod"
-    assert errors[0]["code"] == CheckoutErrorCode.SHIPPING_METHOD_NOT_APPLICABLE.name
-    assert checkout.shipping_method is None
-    assert (
-        mock_is_shipping_method_available.call_count
-        == shipping_models.ShippingMethod.objects.count()
-    )
-
-
-def test_checkout_delivery_method_update_unavailable_variant(
-    staff_api_client,
-    shipping_method,
-    checkout_with_item,
-    address,
-):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    checkout_with_item.lines.first().variant.channel_listings.filter(
-        channel=checkout_with_item.channel
-    ).delete()
-    query = MUTATION_UPDATE_SHIPPING_METHOD
-
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-
-    response = staff_api_client.post_graphql(
-        query,
-        {"id": to_global_id_or_none(checkout_with_item), "shippingMethodId": method_id},
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-
-    checkout.refresh_from_db()
-
-    errors = data["errors"]
-    assert len(errors) == 1
-    assert errors[0]["field"] == "lines"
-    assert errors[0]["code"] == CheckoutErrorCode.UNAVAILABLE_VARIANT_IN_CHANNEL.name
-
-
-@patch("saleor.shipping.postal_codes.is_shipping_method_applicable_for_postal_code")
-def test_checkout_delivery_method_update_excluded_postal_code(
-    mock_is_shipping_method_available,
-    staff_api_client,
-    shipping_method,
-    checkout_with_item,
-    address,
-):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_DELIVERY_METHOD
-    mock_is_shipping_method_available.return_value = False
-
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-
-    response = staff_api_client.post_graphql(
-        query, {"id": to_global_id_or_none(checkout), "deliveryMethodId": method_id}
-    )
-    data = get_graphql_content(response)["data"]["checkoutDeliveryMethodUpdate"]
-
-    checkout.refresh_from_db()
-
-    errors = data["errors"]
-    assert len(errors) == 1
-    assert errors[0]["field"] == "deliveryMethodId"
-    assert errors[0]["code"] == CheckoutErrorCode.DELIVERY_METHOD_NOT_APPLICABLE.name
-    assert checkout.shipping_method is None
-    assert (
-        mock_is_shipping_method_available.call_count
-        == shipping_models.ShippingMethod.objects.count()
-    )
-
-
-# Deprecated
-def test_checkout_shipping_method_update_shipping_zone_without_channel(
-    staff_api_client,
-    shipping_method,
-    checkout_with_item,
-    address,
-):
-    shipping_method.shipping_zone.channels.clear()
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_SHIPPING_METHOD
-
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-
-    response = staff_api_client.post_graphql(
-        query,
-        {"id": to_global_id_or_none(checkout_with_item), "shippingMethodId": method_id},
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-
-    checkout.refresh_from_db()
-
-    errors = data["errors"]
-    assert len(errors) == 1
-    assert errors[0]["field"] == "shippingMethod"
-    assert errors[0]["code"] == CheckoutErrorCode.SHIPPING_METHOD_NOT_APPLICABLE.name
-    assert checkout.shipping_method is None
-
-
-def test_checkout_delivery_method_update_shipping_zone_without_channel(
-    api_client,
-    shipping_method,
-    checkout_with_item,
-    address,
-):
-    shipping_method.shipping_zone.channels.clear()
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_DELIVERY_METHOD
-
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-
-    response = api_client.post_graphql(
-        query, {"id": to_global_id_or_none(checkout), "deliveryMethodId": method_id}
-    )
-    data = get_graphql_content(response)["data"]["checkoutDeliveryMethodUpdate"]
-
-    checkout.refresh_from_db()
-
-    errors = data["errors"]
-    assert len(errors) == 1
-    assert errors[0]["field"] == "deliveryMethodId"
-    assert errors[0]["code"] == CheckoutErrorCode.DELIVERY_METHOD_NOT_APPLICABLE.name
-    assert checkout.shipping_method is None
-
-
-# Deprecated
-def test_checkout_shipping_method_update_shipping_zone_with_channel(
-    staff_api_client,
-    shipping_method,
-    checkout_with_item,
-    address,
-):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_SHIPPING_METHOD
-
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-
-    response = staff_api_client.post_graphql(
-        query,
-        {"id": to_global_id_or_none(checkout_with_item), "shippingMethodId": method_id},
-    )
-    data = get_graphql_content(response)["data"]["checkoutShippingMethodUpdate"]
-
-    checkout.refresh_from_db()
-
-    checkout.refresh_from_db()
-    errors = data["errors"]
-    assert not errors
-    assert data["checkout"]["token"] == str(checkout_with_item.token)
-
-    assert checkout.shipping_method == shipping_method
-
-
-def test_checkout_delivery_method_update_shipping_zone_with_channel(
-    staff_api_client,
-    shipping_method,
-    checkout_with_item,
-    address,
-):
-    checkout = checkout_with_item
-    checkout.shipping_address = address
-    checkout.save(update_fields=["shipping_address"])
-    query = MUTATION_UPDATE_DELIVERY_METHOD
-
-    method_id = graphene.Node.to_global_id("ShippingMethod", shipping_method.id)
-
-    response = staff_api_client.post_graphql(
-        query, {"id": to_global_id_or_none(checkout), "deliveryMethodId": method_id}
-    )
-    data = get_graphql_content(response)["data"]["checkoutDeliveryMethodUpdate"]
-
-    checkout.refresh_from_db()
-
-    checkout.refresh_from_db()
-    errors = data["errors"]
-    assert not errors
-
-    assert checkout.shipping_method == shipping_method
 
 
 def test_query_checkouts(
